cleanup and comments from review
