+++ yarn-project/cli-wallet/src/cmds/deploy_account.ts
@@ -47,6 +47,13 @@ export async function deployAccount(
   };
 
   if (feeOpts.estimateOnly) {
+    /*
+     * This is usually handled by accountManager.deploy(), but we're accessing the lower
+     * level method to get the gas estimates. That means we have to replicate some of the logic here.
+     * In case we're deploying our own account, we need to hijack the payment method for the fee,
+     * wrapping it in the one that will make use of the freshly deployed account's
+     * entrypoint. For reference, see aztec.js/src/account_manager.ts:deploy()
+     */
     const fee =
       !deployOpts?.deployWallet && deployOpts?.fee
         ? { ...deployOpts.fee, paymentMethod: await account.getSelfPaymentMethod(deployOpts.fee.paymentMethod) }
