+++ yarn-project/end-to-end/src/e2e_token_contract/private_transfer_recursion.test.ts
@@ -1,4 +1,3 @@
-import { BatchCall } from '@aztec/aztec.js';
 import { TokenContract, type Transfer } from '@aztec/noir-contracts.js/Token';
 
 import { mintNotes } from '../fixtures/token_utils.js';
