+++ yarn-project/end-to-end/src/fixtures/token_utils.ts
@@ -51,6 +51,7 @@ export async function mintNotes(
   noteAmounts: bigint[],
 ): Promise<bigint> {
   // We can only mint 4 notes at a time, since that's the maximum number of calls our entrypoints allow
+  // TODO(#13024): mint as many notes as possible in a single tx
   const notesPerIteration = 4;
   for (let mintedNotes = 0; mintedNotes < noteAmounts.length; mintedNotes += notesPerIteration) {
     const toMint = noteAmounts.slice(mintedNotes, mintedNotes + notesPerIteration);
